+++ src/test_GLM.cpp
@@ -80,7 +80,7 @@ Rcpp::List test_GLM(const std::string method,
   }
 
   Rcpp::List result = Rcpp::List::create(
-      Rcpp::Named("parTests") = Rcpp::List::create(
+      Rcpp::Named("par_tests") = Rcpp::List::create(
           Rcpp::Named("statistic") = chisq_val,
           Rcpp::Named("convergence") = par_conv),
       Rcpp::Named("optim") = Rcpp::List::create(