+++ NEWS.md
@@ -28,7 +28,9 @@
 
 * `probit` link produces more accurate result in `el_glm()`
 
-* `print` method for an `EL` object shows whether the data are weighted or not.
+* `print()` method for an `EL` object shows whether the data are weighted or not.
+
+* All row or column names (if any) of input data are preserved in a fitted `EL` object.
 
 ### BUG FIXES
 * `confint()` and `confreg()` check if the `cv` argument is compatible with the `th` value set by `control_el()`.